django-blog/requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
certifi==2020.6.20
Django==3.0.8
django-crispy-forms==1.9.2
django-tinymce4-lite==1.8.0
jsmin==2.2.2
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
wincertstore==0.2
